+++ go.mod
@@ -47,11 +47,13 @@ require (
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 	gotest.tools v2.2.0+incompatible
+	helm.sh/helm v2.17.0+incompatible
 	helm.sh/helm/v3 v3.6.1
 	k8s.io/api v0.21.2
 	k8s.io/apimachinery v0.21.2
 	k8s.io/cli-runtime v0.21.0 // indirect
 	k8s.io/client-go v0.21.2
+	k8s.io/helm v2.17.0+incompatible // indirect
 	k8s.io/kubectl v0.21.0 // indirect
 	sigs.k8s.io/controller-runtime v0.9.2
 )
